StudyPulse/requirements.txt
pandas
plotly.express
streamlit
numpy
scikit-learn
transformers
nltk
